Save line ending type in the database.
Creates a "dialect" table and stores the "lineterminator" key and its
value. Not using the csv.Sniffer(), since it has been very unreliable
about detecting the delimiter and line endings. When converting back to
CSV using sqlite2csv, it reads the dialect options from the dialect
table in the database.